Removed Port Forwarding Double Check
Due to an unknown @fsockopen bug (maybe error 111), this caused the script to display a "closed" status which forced the script to kill the VPN connection.

NOTICE: For now, the double check is off. Although, for those running repair though the task scheduler every XX minutes, it will still check the port forwarding status.